update to QT/PySide2 5.15.1 (i don't want to use 5.15.2 because of this https://bugreports.qt.io/browse/PYSIDE-1473) - split Port.hpp / Port.cpp into 3 different files - make Port::clone a pure virtual function as this is now hopefully supported by PySide2 - fix the PySide2 .abi3.so shared library dependencies on linux for - use the PySide2 installation directory for binary shared library dependencies (libicu) of Qt - use new QVideoFrame::image instead of private API function
