+++ nexxT/core/qrc_resources.py
@@ -1,6 +1,6 @@
 # Resource object code (Python 3)
 # Created by: object code
-# Created by: The Resource Compiler for Qt version 5.14.0
+# Created by: The Resource Compiler for Qt version 5.15.1
 # WARNING! All changes made in this file will be lost!
 
 from PySide2 import QtCore
@@ -264,9 +264,9 @@ qt_resource_struct = b"\
 \x00\x00\x00\x00\x00\x02\x00\x00\x00\x02\x00\x00\x00\x02\
 \x00\x00\x00\x00\x00\x00\x00\x00\
 \x00\x00\x00\x10\x00\x00\x00\x00\x00\x01\x00\x00\x00\x00\
-\x00\x00\x01u_\x80\xe6\xc4\
+\x00\x00\x01v\xc7b\x82\xe4\
 \x00\x00\x006\x00\x01\x00\x00\x00\x01\x00\x00\x08h\
-\x00\x00\x01u_g\x0c\x08\
+\x00\x00\x01v\xc7b\x82\xe4\
 "
 
 def qInitResources():
